pyproject.toml
[tool.poetry]
name = "iudex"
version = "0.13.2"
description = ""
authors = ["Drake Wong <drake@iudex.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,!=3.9.7"
opentelemetry-exporter-otlp-proto-http = "^1.25.0"
opentelemetry-distro = "^0.46b0"
wrapt = "^1.16.0"
opentelemetry-semantic-conventions-ai = "^0.3.1"
tiktoken = ">=0.3.3,<1"
# opentelemetry-contrib-instrumentation
# should just be the main package, but one dep is broken: https://github.com/open-telemetry/opentelemetry-python-contrib/issues/2053
opentelemetry-instrumentation-aio-pika = "^0.46b0"
opentelemetry-instrumentation-aiohttp-client = "^0.46b0"
opentelemetry-instrumentation-aiopg = "^0.46b0"
opentelemetry-instrumentation-asyncio = "^0.46b0"
opentelemetry-instrumentation-asyncpg = "^0.46b0"
opentelemetry-instrumentation-aws-lambda = "^0.46b0"
opentelemetry-instrumentation-boto = "^0.46b0"
opentelemetry-instrumentation-boto3sqs = "^0.46b0"
opentelemetry-instrumentation-botocore = "^0.46b0"
opentelemetry-instrumentation-cassandra = "^0.46b0"
opentelemetry-instrumentation-celery = "^0.46b0"
opentelemetry-instrumentation-confluent-kafka = "^0.46b0"
opentelemetry-instrumentation-dbapi = "^0.46b0"
opentelemetry-instrumentation-django = "^0.46b0"
opentelemetry-instrumentation-elasticsearch = "^0.46b0"
opentelemetry-instrumentation-falcon = "^0.46b0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-instrumentation-flask = "^0.46b0"
opentelemetry-instrumentation-grpc = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
opentelemetry-instrumentation-jinja2 = "^0.46b0"
opentelemetry-instrumentation-kafka-python = "^0.46b0"
opentelemetry-instrumentation-logging = "^0.46b0"
opentelemetry-instrumentation-mysql = "^0.46b0"
opentelemetry-instrumentation-mysqlclient = "^0.46b0"
opentelemetry-instrumentation-pika = "^0.46b0"
opentelemetry-instrumentation-psycopg = "^0.46b0"
opentelemetry-instrumentation-psycopg2 = "^0.46b0"
opentelemetry-instrumentation-pymemcache = "^0.46b0"
opentelemetry-instrumentation-pymongo = "^0.46b0"
opentelemetry-instrumentation-pymysql = "^0.46b0"
opentelemetry-instrumentation-pyramid = "^0.46b0"
opentelemetry-instrumentation-redis = "^0.46b0"
opentelemetry-instrumentation-remoulade = "^0.46b0"
opentelemetry-instrumentation-requests = "^0.46b0"
opentelemetry-instrumentation-sqlalchemy = "^0.46b0"
opentelemetry-instrumentation-sqlite3 = "^0.46b0"
opentelemetry-instrumentation-system-metrics = "^0.46b0"
opentelemetry-instrumentation-threading = "^0.46b0"
opentelemetry-instrumentation-tornado = "^0.46b0"
opentelemetry-instrumentation-tortoiseorm = "^0.46b0"
opentelemetry-instrumentation-urllib = "^0.46b0"
opentelemetry-instrumentation-urllib3 = "^0.46b0"
# openllmetry
opentelemetry-instrumentation-alephalpha = "^0.23.0"
opentelemetry-instrumentation-anthropic = "^0.23.0"
opentelemetry-instrumentation-bedrock = "^0.23.0"
opentelemetry-instrumentation-chromadb = "^0.23.0"
opentelemetry-instrumentation-cohere = "^0.23.0"
opentelemetry-instrumentation-google-generativeai = "^0.23.0"
opentelemetry-instrumentation-haystack = "^0.23.0"
opentelemetry-instrumentation-langchain = "^0.23.0"
opentelemetry-instrumentation-llamaindex = "^0.23.0"
opentelemetry-instrumentation-milvus = "^0.23.0"
opentelemetry-instrumentation-mistralai = "^0.23.0"
opentelemetry-instrumentation-ollama = "^0.23.0"
opentelemetry-instrumentation-pinecone = "^0.23.0"
opentelemetry-instrumentation-qdrant = "^0.23.0"
opentelemetry-instrumentation-replicate = "^0.23.0"
opentelemetry-instrumentation-together = "^0.23.0"
opentelemetry-instrumentation-transformers = "^0.23.0"
opentelemetry-instrumentation-vertexai = "^0.23.0"
opentelemetry-instrumentation-watsonx = "^0.23.0"
opentelemetry-instrumentation-weaviate = "^0.23.0"
openai = "1.42.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
openai = "^1.33.0"
fastapi = {extras = ["all"], version = "^0.111.0"}
supabase = "^2.5.1"
modal = "^0.63.12"
guardrails-ai = "^0.5.4"
streamlit = "^1.37.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.pyright]
venvPath = "."
venv = ".venv"
